pyproject.toml
[tool.poetry]
name = "summary-discord-bot"
version = "1.1.0"
description = "A Discord server summary bot"
authors = [
    "Liam <11690542+liammcnamara@users.noreply.github.com>",
    "davidpirogov <580286+davidpirogov@users.noreply.github.com>",
    "William Parry <williamparry@gmail.com>"
]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
discord = "^2.3.2"
better-exceptions = "^0.3.3"
dpn-pyutils = "^3.12.3"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
jinja2 = "^3.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
